requirements.txt

pdfplumber
python-dotenv
streamlit
python-docx
google-generativeai
python-pptx